+++ docs/source/api_ref_utilities.rst
@@ -73,8 +73,6 @@ Utilities to reduce memory consumption during training.
     register_optim_in_bwd_hooks
 
 
-.. _perf_profiling_label:
-
 Performance and Profiling
 -------------------------
 
@@ -87,7 +85,6 @@ of your finetuning job.
 
     get_memory_stats
     log_memory_stats
-    setup_torch_profiler
 
 .. _metric_logging_label:
