+++ recipes/configs/qwen2/7B_lora.yaml
@@ -87,7 +87,7 @@ enable_activation_checkpointing: False
 # Show case the usage of pytorch profiler
 # Set enabled to False as it's only needed for debugging training
 profiler:
-  _component_: torchtune.utils.setup_torch_profiler
+  _component_: torchtune.training.setup_torch_profiler
 
   enabled: False
